feat 장바구니 수정 및 결제기능 구현 진행중
subject : 장바구니 수정 및 결제기능 구현 진행중

+++ So1omon/src/main/java/com/kh/so1omon/board/controller/BoardController.java
@@ -458,7 +458,7 @@ public class BoardController {
     }
     
     @RequestMapping("tboardEnroll.bo")
-    public String insertTboard(TBoard t, int userNo ,Attachment at, MultipartFile thumbnailFile, MultipartFile[] detailFiles ,HttpSession session, Model model) {
+    public String insertTboard(TBoard t, long userNo ,Attachment at, MultipartFile thumbnailFile, MultipartFile[] detailFiles ,HttpSession session, Model model) {
     	
     	
     	
@@ -780,7 +780,7 @@ public class BoardController {
 	// 좋아요
     @ResponseBody
     @RequestMapping("likeAdDel.li")
-    public String likeAdDel(Like l,String boardNo,int userNo) {
+    public String likeAdDel(Like l,String boardNo,long userNo) {
     	
     	System.out.println("boardNo:"+boardNo);
     	System.out.println("userNo:"+userNo);
